requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
cryptography==41.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0